jira_utils/pyproject.toml
[tool.poetry]
name = "jira_utils"
version = "0.1.0"
description = ""
authors = ["mmm1513 <ttugendhaft@grubhub.com>"]

[project]
name = "jira_utils"
description = "Jira Utils"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
jira = "^3.2.0"
PyYAML = "^6.0"
dacite = "^1.6.0"
inquirer = "^2.9.2"

[tool.poetry.dev-dependencies]
ipdb = "^0.13.9"
ipython = "^8.3.0"

[build-system]
# Minimum requirements for the build system to execute.
requires = ["poetry-core>=1.0.0", "setuptools", "wheel"] # PEP 508 specifications.
build-backend = "setuptools.build_meta"
# build-backend = "poetry.core.masonry.api"
